refactor!: move useActionsShowDetails to the admin-settings app
