+++ packages/web-app-admin-settings/src/components/Groups/CreateGroupModal.vue
@@ -19,7 +19,7 @@ import { useGettext } from 'vue3-gettext'
 import { computed, defineComponent, ref, PropType, unref, watch } from 'vue'
 import { Group } from '@opencloud-eu/web-client/graph/generated'
 import { MaybeRef, Modal, useClientService, useMessages } from '@opencloud-eu/web-pkg'
-import { useGroupSettingsStore } from '../../composables'
+import { useGroupSettingsStore } from '../../composables/stores/groupSettings'
 
 export default defineComponent({
   name: 'CreateGroupModal',
