+++ packages/web-app-admin-settings/src/composables/actions/index.ts
@@ -1,2 +1,4 @@
 export * from './groups'
 export * from './users'
+
+export * from './useActionsShowDetails'
