+++ packages/web-pkg/src/composables/actions/index.ts
@@ -3,6 +3,5 @@ export * from './files'
 export * from './spaces'
 export * from './types'
 
-export * from './useActionsShowDetails'
 export * from './useOpenEmptyEditor'
 export * from './useWindowOpen'
